Add alias to duplicate symbols for kallsyms
In the Linux kernel, it is not uncommon for drivers or modules related to
similar peripherals to have symbols with the exact same name.
While this is not a problem for the kernel's binary itself, it becomes an
issue when attempting to trace or probe specific functions using
infrastructure like ftrace or kprobe.

The tracing subsystem relies on the `nm -n vmlinux` output, which provides
symbol information from the kernel's ELF binary. However, when multiple
symbols share the same name, the standard nm output does not differentiate
between them. This can lead to confusion and difficulty when trying to
probe the intended symbol.

**kas_alias** addresses this challenge by extending the symbol names with
unique suffixes during the kernel build process.
The newly created aliases for these duplicated symbols are unique names that
can be fed to the ftrace sysfs interface. By doing so, it enables
previously unreachable symbols to me probed.

Signed-off-by: Alessandro Carminati <alessandro.carminati@gmail.com>